python_backend/requirements.txt
# FastAPI Backend Dependencies
# UptimeAI - Predictive Maintenance Dashboard

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database (SQLite for development, PostgreSQL optional)
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pyotp==2.9.0
qrcode==7.4.2

# Google OAuth
google-auth==2.27.0
google-auth-oauthlib==1.2.0
httpx==0.26.0

# Email
aiosmtplib==3.0.1
email-validator==2.1.0.post1

# Machine Learning
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2

# Utilities
python-dotenv==1.0.1
pydantic[email]==2.6.0
pydantic-settings==2.1.0

# Security
itsdangerous==2.1.2
slowapi==0.1.9

# CORS
starlette==0.35.1

# Production Server
gunicorn==21.2.0

# Database (PostgreSQL support)
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Logging
python-json-logger==2.0.7